+++ src/main/java/com/mailler/controller/sqs/consumer/SqsEmailConsumer.java
@@ -1,4 +1,4 @@
-package com.mailler.controller;
+package com.mailler.controller.sqs.consumer;
 
 import javax.mail.MessagingException;
 
@@ -6,7 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.stereotype.Component;
 
-import com.mailler.controller.sender.EmailSender;
+import com.mailler.controller.ses.sender.Email;
+import com.mailler.controller.ses.sender.EmailSender;
 
 @Component
 public class SqsEmailConsumer {
